refactor: экскурсии создаются правильно

+++ src/main/java/excopen/backend/dto/CoordinateDTO.java
@@ -13,5 +13,5 @@ public class CoordinateDTO {
     private Double longitude;
 
     @NotNull
-    private Integer zoom;
+    private Double zoom;
 }
